Add accurate GPLv3 Copyright & Licensing notice
Add the accurate http://www.gnu.org/licenses/gpl-3.0.md[GPLv3 GNU
General Public License Version 3] Copyright & Licensing notice.

== Problem

Explained with the words of the full GPLv3 license:

  If you develop a new program, and you want it to be of the greatest
  possible use to the public, the best way to achieve this is to make it
  free software which everyone can redistribute and change under these
  terms.

== Solution

Add the accurate GPLv3 Copyright & Licensing notice near the top of a
document as *recommended* by the full GPLv3 license:

  To do so, attach the following notices to the program. It is safest to
  attach them to the start of each source file to most effectively state
  the exclusion of warranty; and each file should have at least the
  "copyright" line and a pointer to where the full notice is found.

.GPLv3 GNU General Public License Version 3
....
<one line to give the program's name and a brief idea of what it does.
Copyright (C) <year>  <name of author>

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.
....

Signed-off-by: Tim Friske <me@tifr.de>