Describe content of files and structure of folders
Describe the content of files and structure of folders to help others to
find their way around in a project.

== Problem

Without a description of the content of files and structure of folders
people willing to use a project or contribute to it will have a hard
time to find what they are looking for.  A missing description also
leaves room for wrong assumptions possibly leading to even worse
decisions as to how to name files and folders and where to place them.
Last but not least a project's content and structure will inevitably
diverge with no common understanding in mind.

== Solution

Spend a few words about where to find what, where to place files and
folders, how to name them and what they should contain.

Signed-off-by: Tim Friske <me@tifr.de>