Explain project's Git workflow
Explain the _Git_ workflow this project follows.

== Problem

Without a thorough description of the _Git_ worfklow employed in a
project contributors are more or less free to branch, commit, merge etc.
whenever and however they like.  Therefore the _Git_ reopsitory will
soon find itself in chaos rendering it useless.  As a consequence the
quality of a project and the collaboration between the team members will
be heavily impaired to the point where people leave the project and the
project might die or live on in another, better organized fork.

== Solution

Describe the _Git_ workflow this project follows in the introductory
documents of a project, e.g. a README file, with pointers where to find
more detailed information.

Signed-off-by: Tim Friske <me@tifr.de>