Move folder license below Documentation
As stated in the file `README.adoc` only the most important or common
files and folders must live in the top-level folder.  But the folder
`license` does not fall into this category.  Therefore move it below the
folder `Documentation` because license information can be seen as a sort
of documentation.

Update the symbolic link `LICENSE.md` as well as all links that point to
the link:Documentation/license/gpl-3.0/gpl-3.0.md[offline version] of
the license GPLv3.

Signed-off-by: Tim Friske <me@tifr.de>