Autostart of GNOME's SSH Agent
Prevent the autostart of _GNOME's_ SSH Agent because it shadows a
running instance of _GnuPG's_ GPG Agent that the user intended to use as
an SSH Agent substitute.  The SSH Agent overwrites the environment
variable `SSH_AUTH_SOCK` with its Unix socket hereby intercepting any
connection attempts to the GPG Agent and subsequent communication with
it.

Signed-off-by: Tim Friske <me@tifr.de>