Code helper for backing up GPG keys + owner trusts
This program should help the user to backup not only GPG keys but also
the trust values the user put in these keys.  The user can either backup
all GPG keys or the ones with the given name/s.

The GPG keys are written in _GnuPG's_ binary format into backup files
which end in `.gpg`.  In case the backup file already exists it is
backed up first by copying it to another backup file by the same name
but extension `.gpg~<N>` where `N` denotes the Nth backup.  The original
backup file — the one ending in `.gpg` — is then overwritten with the
latest backup of the GPG key/s.

Last but not least the program backs up the owner trust values which
should be considered as equally important as the GPG keys themselves.

Signed-off-by: Tim Friske <me@tifr.de>