Advise to keep certification-only GPG keys offline
The private part of a certification-only _GnuPG_ key is the most
sensible and thus precious information that must be kept secret under
all circumstances.  Leaking this private key material is the worst kind
of security breach that can happen as it would not only make this
primary key untrustworthy and thus unusable but all of its secondary
keys too.  In that case an entirely new identity has to be established
and all previously lost trust must be regained over a long period of
time.

Signed-off-by: Tim Friske <me@tifr.de>