Trust GPG keys not present for security reasons
Storing a secret key offline on an encrypted medium raises the level of
security but makes it impossible for _GPG_ to use it to verify the
validity of a given recipient's or signator's key.  In order to make
such verifications possible tell _GPG_ about the secret keys it should
ultimately trust as if they were present on the secret key ring.

Signed-off-by: Tim Friske <me@tifr.de>