Generate Ed25519 SSH key with extra secure options
Generate a modern _SSH_ key that makes use of the _Edwards Curve 25519_
elliptic curve algorithm.

Store the _Ed25519_ _SSH_ key under a non-standard folder and name in
order to reduce the need for overwriting keys generated with standard
options.  This non-standard location should also be a sign that the user
who generated this _Ed25519_ _SSH_ key carefully thought about the key's
purpose and security.

Since this _Ed25519_ _SSH_ key is meant to be used by a real user place
her/his user name as a comment into the generated key.  This helps in
identifying the owner in case s/he cannot be determined via the file's
owner.

_OpenSSH_ already stores _Ed25519_ keys by default in the new and safer
key format.  Therefore there is no need to explicitly pass the option
flag `o` to the `ssh-keygen` program.

In order to make it even harder to brute force the passphrase with which
the _Ed25519_ _SSH_ key is protected increase the KDF key derivation
function rounds to a reasonable number.  Increasing that number means
that also the time it takes to decrypt this _SSH_ key is raised.
Decryption should be `slow' enough to severely hinder brut force attacks
but fast enough to not become an annoyance to the key's owner who knows
the correct passphrase.

Signed-off-by: Tim Friske <me@tifr.de>