Generate RSA-4096 SSH key with extra sec. options
Generate a backwards compatible _SSH_ key that makes use of the _Rivest,
Shamir and Adleman_ public key algorithm with a key size of 4096 bit.

Store the _RSA-4096_ _SSH_ key under a non-standard folder and name in
order to reduce the need for overwriting keys generated with standard
options.  This non-standard location should also be a sign that the user
who generated this _RSA-4096_ _SSH_ key carefully thought about the
key's purpose and security.

Since this _RSA-4096_ _SSH_ key is meant to be used by a real user place
her/his user name as a comment into the generated key.  This helps in
identifying the owner in case s/he cannot be determined via the file's
owner.

_OpenSSH_ does not store _RSA_ keys by default in the new and safer key
format.  Therefore explicitly enable the new key format for this
_RSA-4096_ _SSH_ key by passing the option flag `o` to the `ssh-keygen`
program.

In order to make it even harder to brute force the passphrase with which
the _RSA-4096_ _SSH_ key is protected increase the KDF key derivation
function rounds to a reasonable number.  Increasing that number means
that also the time it takes to decrypt this _SSH_ key is raised.
Decryption should be `slow' enough to severely hinder brut force attacks
but fast enough to not become an annoyance to the key's owner who knows
the correct passphrase.

Signed-off-by: Tim Friske <me@tifr.de>