Generate Ed25519 SSH key with extra secure options
Generate a modern _SSH_ key that makes use of the _Edwards Curve 25519_
elliptic curve algorithm.

Store the _Ed25519_ _SSH_ key under a non-standard folder and name in
order to reduce the need for overwriting keys generated with standard
options.  This non-standard location should also be a sign that the
administrator of a host who generated this _Ed25519_ _SSH_ key carefully
thought about the key's purpose and security.

Since this _Ed25519_ _SSH_ key is meant to be used as a host key place
its host name as a comment into the generated key.  This helps in
identifying the host in case its name cannot be determined elsewise.

_OpenSSH_ already stores _Ed25519_ keys by default in the newer and
safer key format.  Therefore there is no need to explicitly pass the
option flag `o` to the `ssh-keygen` program.

Because an _OpenSSH_ host key cannot be protected by a passphrase it
should be secured at least on file system level.  It is recommended to
restrict the file's owner, owning group and permissions of the key file
holding the private part of that key as much as possible following the
security principle of `least privilige'.  Here the owner is the user
`root', the owning group the group `root' and only the owner permitted
to read the private key file.

Signed-off-by: Tim Friske <me@tifr.de>