Ask user to choose SSH RSA-4096 key comment
Let the user accept the default key comment or her/him choose her/his
favorite key comment.

The previous design decision to choose the host name as a key comment
was a bad choice because it didn't reveal much about the purpose of the
_SSH_ key.  Also the previous implementation didn't let the user choose
a comment of her/his choice.

The current implmentation has the following advantages over the old one:

  . The user is prompted to either accept the default key comment or
    overwrite it with her/his favorite key comment.

  . The default key comment provides the following vital information:

    .. The type of _SSH_ key: RSA-4096.

    .. The host for which the _SSH_ key was generated.

    .. The creation date of the _SSH_ key.

Signed-off-by: Tim Friske <me@tifr.de>