Rephrase rationale behind SSH key gen comment
The previous statement about embedding the host name as the sole comment
of a _SSH_ key doesn't hold anymore because that is way too little info
for an entire comment.

Instead embed a given comment that should identify the _SSH_ key and
make it possible to differentiate this key from others.

Signed-off-by: Tim Friske <me@tifr.de>