Restrict access to SSH host key files
Only the owner must have the permission to read and write her/his _SSH_
key files; in case of the host keys usually the user `root'.  The key
file containing the public key of the public/private key pair may
additionally be read by the group and others because it comprises only
public information that must be shared in order for public key
cryptography to work.

Signed-off-by: Tim Friske <me@tifr.de>