Clarify structure of OpenSSH client configuration
The previous explanation did not start with an introduction to the
overall structure but jumped right into the details of explaining that
only the first occurrence of any given setting becomes effective.  Also
that explanation was inaccurate because it claimed that only `Host`
expressions would start a new section; `Match` expressions start new
sections too.

The rewritten explanation gives an overview about the following points
starting from a bird's-eye view in the following order:

  . The configuration file may be divided into sections.

  . Which keywords start a section.  Where does a section end.

  . Sections define settings specific to those.

  . The order in which sections are defined is important.  More
    specific sections must come before less sepcific ones.

  . In the gobal section fallback settings may be defined.  This section
    must come after all other sections at the end of file.

Signed-off-by: Tim Friske <me@tifr.de>