Prevent root from logging in via OpenSSH daemon
Don't rely on _OpenSSH's_ daemon to allow the super user `root' to login
by default.  Such logins may breach not only the security of the account
of the user `root' but worse the security of the host or even the entire
network because s/he is the super user with all privileges one can think
of.

Signed-off-by: Tim Friske <me@tifr.de>