Achieve min authentication tries of OpenSSH client
When the _OpenSSH_ client is asked by a _SSH_ daemon to authenticate
itself via public key authentication the client firstly tries out all
private keys that are offered by a _SSH_ agent and secondly all those
that are stored on the file system.

This trial-and-error method starts to cause problems when the number of
_SSH_ keys becomes larger.

With each wrong key offered during authentication there will be one less
attempt left until the limit of a _SSH_ daemon is reached and the entire
authentication procedure fails.

Without _OpenSSH's_ connection multiplexing feature it takes fairly long
until the client established a secure connection to a daemon.  Even
worse the total time it takes increases with each key that the client
offers but a daemon rejects.

Setting the `IdentitiesOnly` option to `yes` causes the _OpenSSH_ client
to only try the keys which are either explicitly given or implicitly
assumed to exist under their default names.  To keep the configuration
simple this setting should be made once in the global section of the
system-wide configuration file of _OpenSSH's_ client.

For the `IdentitiesOnly` option to lead to a noticeable reduction in the
number of keys tried for authentication there must be either only one
`IndentityFile` setting in the global section of the user configuration
file of _OpenSSH's_ client or one per section for a host or group of
hosts.  Therefore the `IdentityFile` settings have to move from the
system- to the user configuration file.

Signed-off-by: Tim Friske <me@tifr.de>