Split IdentitiesOnly and AddKeysToAgent settings
Logically separate these _OpenSSH_ client settings because they are
truly independent from each other.  Jointly configuring them by setting
their values in unison does not add any value.

Also rephrase the reasoning behind setting the `AddKeysToAgent` option
to `yes`.  It is not true anymore that `each private key read from an
existing file listed below' is added to a _SSH_ agent because the
`IdentityFile` settings were removed from the system-wide configuration
file of _OpenSSH's_ client.  Instead they must be made on a host-by-host
basis in the user configuration file.

The previous description of the `AddKeysToAgent` setting was too
detailed.  The reader doesn't care to which particular implementation of
a _SSH_ agent the user's keys are added.

Signed-off-by: Tim Friske <me@tifr.de>