Conf Systemd socket FileDescName for GnuPG Agent
The `FileDescriptorName` option assigns a name to all file descriptors
this socket unit encapsulates.  This is useful to help activated
services identify specific file descriptors if multiple ones are passed.

If this setting is not used the file descriptor name defaults to the
name of the socket unit file including its `.socket` file extension.

Signed-off-by: Tim Friske <me@tifr.de>