Conf Systemd socket Service for GnuPG Agent
The service unit must be explicitly specfied by this socket unit because
their file names, ignoring file extensions, differ.

The `Service` option configures the name of the service unit to start
when activated through the socket unit.

If this setting is not used the name of the service unit file defaults
to the name of the socket unit file ignoring their file extensions
`.service` and `.socket` respectively.

Signed-off-by: Tim Friske <me@tifr.de>