Add global configuration file for Git
This file shall contain the _Git_ settings common to all _Git_
repositories of the user unless overruled by the configuration of
certain repositories.  It should be copied into the folder
`$XDG_CONFIG_HOME/git`, given the name `global.conf` and symbolically
linked under the name `config` in the same folder in order for _Git_ to
find it.

Signed-off-by: Tim Friske <me@tifr.de>