Run Git Log's alias `lf` in current directory
The intention is to let _Git_ _Log's_ alias `lf` operate on files and
folders below the current directory.

Previously when the user specified the path `.` to refer to the current
directory _Git_ always resolved it as the top-level directory of the
repository even though the user ran the aliased _Git_ command from a
subdirectory.  This is what the user doesn't expect most likely.

But according to _Git's_ manual page man:git-config[1] _Git_ treats
aliases starting with an `!` exclamation mark as shell commands which it
runs from the top-level directory of a repository.  This is feasible for
scripted executions of _Git_ commands but impractical for interactive
use.

Signed-off-by: Tim Friske <me@tifr.de>