Add global config file for misc settings of Git
This file shall define the _Git_ variables pertaining to miscellaneous
aspects of _Git_ for which an own configuration file does not exist yet
and a new one would be too much.  _Git_ should be customized here to the
user's needs such that s/he can accomplish her/his day-to-day tasks with
ease.

It is recommended to copy this file into the folder `global.d` below
_Git's_ configuration folder `$XDG_CONFIG_HOME/git`, give it the name
`misc.conf` and include it in _Git's_ global configuration file
`$XDG_CONFIG_HOME/git/config` (symbolic link to `global.conf`).

Signed-off-by: Tim Friske <me@tifr.de>