Include Git's global config file of misc settings
The customizations made to miscellaneous aspects of _Git_ should not be
embedded in _Git's_ global configuration file but externalized and
included in order to give this file a structure that is easy to
understand and extend.  The overall idea is it to split _Git's_ global
configuration file into logical pieces all of which are defined in their
own files and included by it.

Signed-off-by: Tim Friske <me@tifr.de>