Develop program to list types of files
Listing the types of files should be as simple as calling `file-type
<file (on path)>...`.  The program shall list the types of all given
files.  If an argument doesn't resolve to a file below the current
directory the program shall perform a path search.  In case a file was
found in one of the directories defined in the `$PATH` environment
variable the program shall report the type of that file instead.

Signed-off-by: Tim Friske <me@tifr.de>