Develop program to manage $PATH environ. variable
Managing the `$PATH` environment variable should be as simple as calling
`path`.  The program shall be able to print the `$PATH` in its `raw' and
`pretty' forms as well as with a path entry either appended, prepended
or removed.  It shall treat the `$PATH` as set in which each path entry
may occur only once or never.

Signed-off-by: Tim Friske <me@tifr.de>