Test if $PATH contains given path entry
Use a pattern in form of a regular expression to test whether the
`$PATH` contains a given path entry.

The prefix `(^|.*:)` ensures that an entry only matches if it either
occurs at the start of `$PATH` or it is separated with the `:` colon
separator from another entry that precedes it.  Likewise the suffix
`($|:.*)` enusres that an entry correctly ends.

Signed-off-by: Tim Friske <me@tifr.de>