Let Ls classify file types with suffix characters
Let the file & folder listing program _Ls_ classify file types with
suffix characters in order to make the file types also visible in
listings other than in long format.  One of the characters from the
string `*/=>@|' is appended to a file system object's name in order to
indicate its type.

Signed-off-by: Tim Friske <me@tifr.de>