Let Bash list names of stopped jobs and don't exit
Enable the `checkjobs` shell option in order to alleviate the user from
having to manually execute the shell builtin `jobs` after an interactive
shell s/he is about to exit told her/him that there are stopped jobs.
When this option is enabled _Bash_ also lists the names of stopped jobs
helping the user to decide if s/he want's to really forcefully or rather
gracefully quit the stopped jobs.

Signed-off-by: Tim Friske <me@tifr.de>