Let Bash actively prevent garbled terminal output
Programs when not informed about the current terminal screen size might
mess up the terminal screen with their output.  Therefore let _Bash_
update the `$LINES` and `$COLUMNS` variables after each command in order
to continuously reflect the current terminal screen dimensions.  For
example the dimensions change when the terminal emulator is resized or
the terminal screen is split by a terminal multiplexer into windows and
panes.

Signed-off-by: Tim Friske <me@tifr.de>