Prevent data loss from accidental file overwrites
_Bash's_ file redirection operator `>` to overwrite files is all to
easily typed altough what the user really meant was to append to a file
with the similar operator `>>`.  By enabling _Bash's_ `noclobber` option
it can be told to inhibit the use of the `>` operator.  Should the user
still wish to overwrite a file s/he must use the `>|` operator which is
harder to type than just `>`.

Signed-off-by: Tim Friske <me@tifr.de>