Define alias to run Ls command w/ default options
Run the _Ls_ command that lists files and directories with default
options which the user is free to configure in one of _Bash's_
user-defined startup file for login shells, e.g. `~/.bash_profile`,
`~/.bash_login` or `~/.profile`.

By making the _Ls_ command configurable the user is not forced to
manually specify her/his favorite options every time s/he invokes it.
This measure should contribute to an overall more pleasant and efficient
day-to-day command line usage because the _Ls_ command is probably one
of the most frequently used commands.

Signed-off-by: Tim Friske <me@tifr.de>