Summarize disk usage up to current directory
The default behavior of the _Du_ command is to report the disk usage of
all directories up to and including the current directory but this is
simply too much information for the user to comprehend.  What the user
usually wants is to list the total disk usage of every directory one
level below and up to the current directory, i.e. without recursively
descending into each and every directory.

Also sort _Du's_ output from highest to lowest disk usage from top to
bottom because the user is naturally interested in the most memory
consuming files and folders.

Signed-off-by: Tim Friske <me@tifr.de>