Dynamically add Bash aliases for all Git aliases
Due to the large number of _Git_ aliases users might have defined it
would be a daunting and error-prone task to redefine every _Git_ alias
as _Bash_ alias.  For this reason dynamically load and programmatically
define one _Bash_ alias for each _Git_ alias declared in any of _Git's_
configuration files.

Signed-off-by: Tim Friske <me@tifr.de>