Store Ls' default options in array varariable
By appending each option to _Ls'_ string environment variable
`$LS_X_OPTIONS` the user could previously join two adjacent options by
accident.  This could easily happen if s/he forgot to separate the
options with whitespace from one another.

A more explicit and less error-prone method is it to define an array
variable and initialize it with an empty array.  Each element is then
appended to the array with the `+=(...)` operation, i.e. without an
index.

Such an array variable can later be easily expanded to a space-separated
string by dereferencing the variable as `"${ARRAY_VARIABLE[@]"`.  In
order to separate the variable with not spaces but arbitrary signs one
has to evaluate it as `"$(IFS=:; echo "${ARRAY_VARIABLE[*]})"`.

Signed-off-by: Tim Friske <me@tifr.de>