Store Grep's default options in array varariable
Using string variables to store multiple strings which are seprated by
whitespace or some other separator sign can easily lead to configuration
errors.  For example the user could join two adjacent strings by
accident, e.g. if s/he forgets to separate the strings with whitespace
from one another.

A more explicit and less error-prone method is it to define an array
variable and initialize it with an empty array.  Each element is then
appended to the array with the `+=(...)` operation, i.e. without an
index.

Such an array variable can later be easily expanded to a space-separated
string by dereferencing the variable as `"${ARRAY_VARIABLE[@]"`.  In
order to separate the variable with not spaces but arbitrary signs one
has to evaluate it as `"$(IFS=:; echo "${ARRAY_VARIABLE[*]})"`.

Signed-off-by: Tim Friske <me@tifr.de>