Store file name under which Mktar program is run
Depending under which name this program is called it shall behave
differenly and create different types of un-/compressed _TAR_ _Tape
Archive_ files.

It is important to remove any leading directories from the program's
file path (zeroth argument, `$0`) because we are only interested in the
program's file name, e.g. for conditional tests.

Signed-off-by: Tim Friske <me@tifr.de>