Let Mktar program print message that it ended
Let the _Mktar_ program print a message that it archived the given file
or directory and the path to the _TAR_ _Tape Archive_ file.  Due to the
variable length of that message the length of the separator line must be
computed such that both lines are equally long before they are printed.

Signed-off-by: Tim Friske <me@tifr.de>