Let Ls list file dates in `long' ISO-8601 format
Let the _Ls_ program display the date-time of files and directories in
the ISO-8601 `long' format.  Due to the format's fixed-width pattern
`+%Y-%m-%d %H:%M` it should be more readable and independent from the
current locale.  With up to the minute precision it should be good
enough for everyday use.

Signed-off-by: Tim Friske <me@tifr.de>