Delete aliases 600, 644, 755 to change file perms
In their current form these aliases are of rather limited use because by
definition of their numeric file modes they are intended to operate on
either files or directories.  Also users usually want to conveniently
fix permissions of an entire tree of files and directories and not on a
file-by-file basis.

Signed-off-by: Tim Friske <me@tifr.de>