Fix cutting names of aliases from Git config files
Call a command and *not* unintentionally some alias or function by the
same name.  This can accidently happen because the code is contained in
a startup file sourced by the same instance of a _Bash_ interactive
shell that might also define an alias or function that shadows a command
by the same name.

Signed-off-by: Tim Friske <me@tifr.de>