Use Tmux with Fzf if `FZF_TMUX` variable is set
The user may set this variable to some value other than `0` zero in
order to signal the fuzzy finder program _Fzf_ that it should use the
terminal multiplexer program _Tmux_ if all other conditions are met.

The environment variable `FZF_TMUX` is usually configured in the startup
file for login shells of the user's shell.

Signed-off-by: Tim Friske <me@tifr.de>