Develop program to run Fzf in appropriate environ
This program shall be invoked by other programs that want to run the
fuzzy finder program _Fzf_ in the best possible environment.  For this
purpose _Fzf_ is invoked with different arguments depending on various
environmental conditions, e.g. whether a terminal multiplexer program
such as _Tmux_ or _Screen_ is already running.

Signed-off-by: Tim Friske <me@tifr.de>