Develop program to change directories with Fzf
With this program the user shall be able to quickly change from the
current directory to one of its child directories with the help of the
fuzzy finder program _Fzf_.

When invoked the user shall be presented with a list of subdirectories
which s/he filters before selecting the directory in question.

The command to find subdirectories shall be externally configurable by
the user.  By default this program shall utilize _GNU's_ _Find_ program.

The command to sort subdirectories shall be externally configurable by
the user.  By default this program shall utilize _GNU's_ _Sort_ program.

Signed-off-by: Tim Friske <me@tifr.de>