Let Fzf-history print usage if conn to terminal
The output of the _Bash's_ `history` built-in command is meant to be
piped into the _Fzf-history_ command because it wouldn't return the
history from the user's interactive shell if directly invoked from a
_Bash_ script such as _Fzf-history_.

_Bash's_ history functionality is only enabled (`set -o history`) for
interactive shells by default.  Enabling this option in a _Bash_ script
also doesn't help to make _Bash's_ `history` built-in command return the
history from the script's parent shell.  That's why this no solution
either.

Signed-off-by: Tim Friske <me@tifr.de>