Tell Fzf-history to prefer more recent commands
Assume that the user usually chooses the more recent of two commands
from history; not the shorter of the two.

The fuzzy finder program _Fzf_ prefers shorter entries first and earlier
entries only then if they have the same length by default.  This default
behavior is reflected in _Fzf's_ `tiebreak` option set to `length,index`
if not specified otherwise.

Signed-off-by: Tim Friske <me@tifr.de>