Let Fzf-history remove index column from command
Remove the index column from the history entry the user selected to only
print the command but not its index in history.  The index column should
only be there when displaying the history but must not be there when
returning a command from it to _Bash's_ editing buffer for execution.

Signed-off-by: Tim Friske <me@tifr.de>