Let Fzf-jobs print usage if connected to terminal
The output of _Bash's_ `jobs` built-in command is meant to be piped into
the _Fzf-jobs_ command because otherwise it wouldn't return the jobs
from the user's interactive shell if directly invoked from a _Bash_
script such as _Fzf-jobs_.

_Bash's_ job control functionality is only enabled (`shopt -o monitor`)
for interactive shells by default.  Enabling this option in a _Bash_
script also doesn't help to make _Bash's_ `jobs` built-in command return
the jobs from the script's parent shell.  That's why this is no solution
either.

The usage instructions advise the user to define aliases shadowing
_Bash's_ built-in commands `jobs`, `bg` and `fg`.  In case the user
follows this advise it becomes very important that s/he escapes these
names with the `\` backslash sign if the intention is to invoke the
shadowed _Bash_ built-in commands.

Signed-off-by: Tim Friske <me@tifr.de>