Develop Readline configuration for Fzf
The file `inputrc.fzf` shall contain _Bash's_ _Readline_ settings
specific to the fuzzy finder program _Fzf_.  Please note that this file
should be read by _Bash's_ startup file for interactive shells
(`~/.bashrc) but only if _Fzf_ is installed and, this file exists and is
readable.  _Readline's_ _Inputrc_ syntax doesn't support such complex
conditions to be tested.  That is why the file `inputrc.fzf` should not
be read by `inputrc` by including it.

Signed-off-by: Tim Friske <me@tifr.de>