Let Fzf-history print help, not usage
The usage information of a program is usually only an abstract of its
command line syntax pointing out the various ways in which that program
can be invoked.  The usage is usually printed as part of a program's
help.  Therefore make use of the term `help' instead of `usage' where
appropriate.

Signed-off-by: Tim Friske <me@tifr.de>