Convert Fzf-history's var `fzf_options` into array
When an array variable is expanded into a string of quoted values the
array elements get seprated by the first character in the `$IFS`
variable which is the space character by default.  Compared to a string
variable with variables to expand and whitespace to separate their
values an array variable has the advantage that a sequence of whitespace
characters is reduced to a single space character upon array expansion.

Signed-off-by: Tim Friske <me@tifr.de>