Let Fzf-cd print help to stderr w/ Cat and Heredoc
Separate _Fzf-cd's_ normal output from its abnormal output such that the
two do not get mixed up and are therefore easier to redirect to distinct
file descriptors.  When invoked without arguments _Fzf-cd_ prints the
_Cd_ command line to the standard output stream but when invoked with
its `help` option it prints its help to the standard error stream.

Use a _Bash_ _Heredoc_ to structure and format text documents.  For this
task it is much better suited than a multitude of subsequent `printf` or
`echo` commands each of which would have to normally precede a line of
some text document.

Signed-off-by: Tim Friske <me@tifr.de>