Develop Fzf-find prog to find file system objects
This program shall allow the user to find file system objects below the
current folder.

When given the `help` option it shall print its purpose, usage, options
and further explanatory text to the standard error stream.  Unless
redirected this stream is connected to the terminal on which the program
is run such that the user can read the help text.

_Fzf-find_ is intended to find only commonly used file system objects
such as files, folders and links.  Therefore it shall assume the
following sensible defaults when run without arguments:

  * Start below the current directory.

  * Search for all kinds of file system objects.

  * Exclude hidden file system objects.

  * Exclude special file systems such as `proc`, `sysfs` etc.

The type of file system objects to include shall be restrictable by the
program calling _Fzf-find_.  Also whether hidden objects should be
included in the search shall be externally determinable.

Since the names of file system objects can contain all kinds of special
characters except the `\0` null character _Fzf-find_ shall be able to
correctly handle such names.  This usually means reading and writing
null-terminated entries when the standard input and output streams are
connected to other programs in the pipe but printing quoted, new-line-
terminated entries for readability when connected to a terminal.

Signed-off-by: Tim Friske <me@tifr.de>