Let Fzf-find handle special chars in file names
Don't let it get confused when special characters appear in the names of
file system objects it finds, e.g. new-line characters.

In case _Fzf-find_ writes to a terminal it quotes the names before it
prints them as new-line-separated entries.  This is done in order to not
mess up the output to the terminal and to keep each entry referrable by
other programs such as _Bash_.

When the standard output stream of _Fzf-find_ is not connected to a
terminal let it write the names as null-terminated entries to that
stream.  This way other programs may also read names that contain
special characters which would be misinterpreted otherwise.

Signed-off-by: Tim Friske <me@tifr.de>