Develop Fzf-sort program to sort records
This program shall read records from the standard input stream, sort
them either alphabetically or ascendingly in the given order and write
them to the standard outout stream in that sort order.  With
_Fzf-sort's_ `reverse` variable set to either `1` or `true` the records
shall be sorted in reverse order.

When given the `help` option _Fzf-find_ shall print its help text and
exit.

This program shall print one record per line and quote it when connected
to a terminal for readability.  But when its output is redirected, e.g.
piped through another program for further processing, it shall print
unquoted but null-terminated records instead to make them securley
processible.

Signed-off-by: Tim Friske <me@tifr.de>