Print Fzf-sort's help as Heredoc to standard error
Since help texts are structured text documents it is best to print them
out with the help of a _Bash_ _Heredoc_ statement.  The alternative is a
multitude of subsequent `echo` and `printf` statements which would lead
to less readable code.

Most programs print their help texts and error messages to the standard
error stream.  This helps to keep the normal program output separate
from its output in case of errors.

Signed-off-by: Tim Friske <me@tifr.de>