Let Fzf-sort handle special characters in records
Don't let it get confused when special characters appear in the records
it sorts, e.g. new-line characters.

In case _Fzf-sort_ writes to a terminal it quotes the records before it
prints them as new-line-separated records.  This is done in order to not
mess up the output to the terminal and to keep each record referrable by
other programs such as _Bash_.

When the standard output stream of _Fzf-sort_ is not connected to a
terminal let it write null-terminated records to that stream.  This way
other programs may also read records that contain special characters
which would be misinterpreted otherwise.

Signed-off-by: Tim Friske <me@tifr.de>